training works fine, but loss doesn't go down

+++ data_generation.py
@@ -3,6 +3,7 @@ import torch
 from utils import RANDOM_SEED
 import numpy as np
 from random import sample
+from config.config import get_config
 def train_test_data_gen(size, number_of_lists, percentage_of_train, seed=RANDOM_SEED):
     '''
     Generate number_of_lists times random lists of unique integers <= size of size 'size'
@@ -22,10 +23,10 @@ def train_test_data_gen(size, number_of_lists, percentage_of_train, seed=RANDOM_
     y_train = [sorted(lst) for lst in X_train]
     y_test = [sorted(lst) for lst in X_test]
 
-    return torch.tensor(X_train, dtype=float, requires_grad=True), \
-            torch.tensor(X_test, dtype=float, requires_grad=True), \
-            torch.tensor(y_train, dtype=float, requires_grad=True), \
-            torch.tensor(y_test, dtype=float, requires_grad=True) 
+    return torch.tensor(X_train, dtype=torch.float32, requires_grad=True), \
+            torch.tensor(X_test, dtype=torch.float32, requires_grad=True), \
+            torch.tensor(y_train, dtype=torch.float32, requires_grad=True), \
+            torch.tensor(y_test, dtype=torch.float32, requires_grad=True)
